# docs/swagger.yaml
basePath: /api
definitions:
...
  commodity.CreateCommodityRequest:
    properties:
      category:
        type: string
      default_name:
        type: string
      produce_address:
        type: string
      produce_at:
        type: string
    type: object
  commodity.SearchQuery:
    properties:
      accurate:
        type: boolean
      range:
        type: string
      search:
        type: string
    type: object
  common.ErrorDetailElement:
    properties:
      field:
        type: string
      message:
        type: string
      param:
        type: string
      struct_field:
        type: string
      tag:
        type: string
      value: {}
    type: object
  common.HttpError:
    properties:
      code:
        type: integer
      detail:
        items:
          $ref: '#/definitions/common.ErrorDetailElement'
        type: array
      message:
        type: string
    type: object
  favorite.AddCommodityFavoriteModel:
    properties:
      commodity_id:
        type: integer
    required:
    - commodity_id
    type: object
  favorite.AddFavoriteModel:
    properties:
      item_id:
        type: integer
    required:
    - item_id
    type: object
  favorite.PriceLimitModel:
    properties:
      item_id:
        type: integer
      price_limit:
        type: number
    type: object
...
  models.Commodity:
    properties:
      category:
        type: string
      default_name:
        type: string
      id:
        type: integer
      produce_address:
        type: string
      produce_at:
        $ref: '#/definitions/models.MyTime'
    type: object
  models.CommodityItem:
    properties:
      commodity:
        $ref: '#/definitions/models.Commodity'
      commodity_id:
        type: integer
      id:
        type: integer
      item_name:
        type: string
      platform:
        $ref: '#/definitions/models.Platform'
      platform_id:
        type: integer
      price:
        type: number
      seller:
        $ref: '#/definitions/models.Seller'
      seller_id:
        type: integer
      update_at:
        $ref: '#/definitions/models.MyTime'
    type: object
  models.Favorite:
    properties:
      commodity_item_id:
        type: integer
      commodityItem:
        $ref: '#/definitions/models.CommodityItem'
      price_limit:
        type: number
      update_at:
        $ref: '#/definitions/models.MyTime'
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
...
  models.MyTime:
    properties:
      time.Time:
        type: string
    type: object
  models.Platform:
    properties:
      country:
        type: string
      id:
        type: integer
      name:
        type: string
      url:
        type: string
    type: object
...
  models.Seller:
    properties:
      address:
        description: Age      int
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
...
  models.User:
    properties:
      age:
        type: integer
      email:
        type: string
      gender:
        type: boolean
      id:
        type: integer
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    email: polaris@fduhole.com
    name: Maintainer Polaris
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Database System Design Project
  version: 1.0.0
paths:
...
  /api/commodities:
    delete:
      consumes:
      - application/json
      description: 删除商品
      parameters:
      - description: commodity id
        in: path
        name: id
        required: true
        type: string
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 删除商品
      tags:
      - 商品
    post:
      consumes:
      - application/json
      description: 添加商品
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/commodity.CreateCommodityRequest'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 添加商品
      tags:
      - 商品
    put:
      consumes:
      - application/json
      description: 更新商品
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Commodity'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 更新商品
      tags:
      - 商品
...
  /api/commodities/search:
    post:
      consumes:
      - application/json
      description: 搜索商品
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/commodity.SearchQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Commodity'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 搜索商品
      tags:
      - 商品
...
  /api/favorites:
    post:
      consumes:
      - application/json
      description: AddFavorite
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/favorite.AddFavoriteModel'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: AddFavorite
      tags:
      - Favorite
  /api/favorites/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteFavorite
      parameters:
      - description: item id
        in: path
        name: id
        required: true
        type: string
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: DeleteFavorite
      tags:
      - Favorite
  /api/favorites/all:
    get:
      consumes:
      - application/json
      description: GetAllFavorites
      parameters:
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Favorite'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: GetAllFavorites
      tags:
      - Favorite
  /api/favorites/commodity:
    post:
      consumes:
      - application/json
      description: AddCommodityFavorite
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/favorite.AddCommodityFavoriteModel'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: AddCommodityFavorite
      tags:
      - Favorite
...
  /api/price/limit:
    post:
      consumes:
      - application/json
      description: AddPriceLimit
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/favorite.PriceLimitModel'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: AddPriceLimit
      tags:
      - Favorite
...
swagger: "2.0"